Script rewritten. 1st attempt.
Now an initrd created for a specific kernel version, will activate rapiddisk only if it was installed for that kernel version. So, if you have kernel version 1.0 and 2.0, and two initrd file, one for kernel version 1.0 and 2.0, and you use the script specifying kernel version 1.0, even if the initrd file for kernel version 2.0 is rebuild, it will NOT contain all the needed parts to activate rapiddisk.

It behaves like dkms: rapiddisk is installed system-wide but can be activated on per-kernel base.