Changed the dkms installing behaviour.
Now the copy operation to /usr/src of the needed files is performed by dkms.
This way, the module is recompiled by dkms immediately, and installed for
the running kernel. Since the modules does not need to be already compiled,
a make clean operation is performed before dkms-install/dkms-uninstall.
Since dkms will install the compiled modules in its dir tree, a make uninstall
operation is invoked too, so that only the modules compiled by dkms will remain
installed in the system. This happens on both dkms-install and dkms-uninstall.

The missing 'all' dependency was added to the install target.

The PWD variable was not appropriate, since was read by make from the shell
and never changed upon recursive make -C invocations. CURDIR variable is an
internal make variable, which is updated to reflect the real current working
directory.