Reworked the Ubuntu install approach to be smarter
Since update-initramfs copies all the boot scripts into the initrd file before
invoking any hook script, it is not possible for rapiddisk's hook script to
customize rapiddisk's boot script directly, with proper parameters such as
ramdisk size etc.
Now, a generic boot script is installed, which during boot, checks for the
existence of a secondary boot script, and executes it when found. If this
secondary script does not exists, the primary boot script does nothing.
The secondary script is created and added to the initrd file by the new
hook script. The secondary script is the one which runs the proper commands
to activate rapiddisk. Parameters such as ramdisk size and root device are
stored in a config file, named after the kernel version it refers to.
Upon invocation, the new hook script generates and add the secondary script
(and all rapiddisk's binaries) only if it finds the config file that matches
the kernel version the initrd file is being build for. This approach prevents
update-initramfs to add rapiddisk's files to initrd files for whom no explicit
install (script's '--install' parameter) action was performed. These initrd
files will contain only the primary boot script, which will do nothing.
On the other hand, an initrd file related to a kernel version for whom an
explicit install action was performed, will end up containing the primary
boot script, the secondary customized boot script and all the needed binaries.
On boot, the primary boot script will find and execute the secondary one,
which will execute rapiddisk.
This approach mimics dracut behaviour on CentOS.