Try to invoke dkms to build missing kernel modules (Ubuntu)
Add a prerm hook script called when a kernel package is removed (Ubuntu
deb package). This way the files related to that kernel are removed too.
Improve the update/remove logic of the deb (Ubuntu deb package). Upon
update, current config is saved and put back in place when the new
package version is installed.